fix(button-group): swap states and interaction states samples
