+++ browser/public/meta.json
@@ -1,1 +1,1 @@
-{"version":"23.2.915","date":"2026-08-12 15:30:01","note":"this file is auto-generated"}
+{"version":"23.2.915","date":"2026-08-14 20:16:59","note":"this file is auto-generated"}
