+++ samples/inputs/button-group/states/package.json
@@ -1,6 +1,6 @@
 {
   "name": "react-button-group-states",
-  "description": "This project provides example of Button Group States using Infragistics React components",
+  "description": "This project provides example of Button Group Interaction States using Infragistics React components",
   "author": "Infragistics",
   "homepage": ".",
   "version": "1.4.0",
